% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accepted_mntaxa.R
\name{accepted_mntaxa}
\alias{accepted_mntaxa}
\title{Create table of accepted taxa}
\usage{
accepted_mntaxa(
  taxonomy_levels = FALSE,
  sources = FALSE,
  phys = FALSE,
  origin = FALSE,
  common = FALSE,
  cvals = FALSE,
  exclude = FALSE
)
}
\arguments{
\item{taxonomy_levels}{Binary option (TRUE, FALSE) to include rank of taxa and taxonomic parents.}

\item{sources}{Binary option (TRUE, FALSE) to include authorities and publications.}

\item{phys}{Binary option (TRUE, FALSE) to include physiognomy.}

\item{origin}{Binary option (TRUE, FALSE) to include origin.}

\item{common}{Binary option (TRUE, FALSE) to include common names.}

\item{cvals}{Binary option (TRUE, FALSE) to include c-values.}

\item{exclude}{Binary option (TRUE, FALSE) to include information about excluding taxa from analyses.}
}
\value{
Tibble of currently accepted taxa in MNTaxa and any additional information requested by the user.
}
\description{
Create table of accepted taxa
}
\examples{
accepted <- accpeted_mntaxa()
}
