% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{auth_raw}
\alias{auth_raw}
\title{Author codes table}
\format{
A data frame with the following columns:
\describe{
\item{id}{Integer. Primary key.}
\item{author}{Character. Abbreviated or full name of the taxonomic authority.}
}
}
\source{
MNTaxa database, plants.authors table.
}
\usage{
auth_raw
}
\description{
Author codes table
}
\keyword{datasets}
