% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_names.R
\name{combine_names}
\alias{combine_names}
\title{Function to combine names in a character string}
\usage{
combine_names(x)
}
\arguments{
\item{x}{A vector of taxon names.}
}
\value{
A character string. If all names in the string are the same genus, the genus name will appear once at the beginning. Otherwise, the string will contain all full names with "/" in between.
}
\description{
Function to combine names in a character string
}
\examples{
combine_names(c("Amelanchier humilis", "Amelanchier sanguinea"))
}
