% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{exclude_codes_raw}
\alias{exclude_codes_raw}
\title{Exclusion reason codes table}
\format{
A data frame with the following columns:
\describe{
\item{id}{Integer. Primary key.}
\item{excluded_reason_code}{Character. Abbreviated code for the exclusion reason.}
\item{description}{Character. Full description of the exclusion reason.}
}
}
\source{
MNTaxa database, plants.excluded_reasons table.
}
\usage{
exclude_codes_raw
}
\description{
Exclusion reason codes table
}
\keyword{datasets}
