% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_mntaxa.R
\name{lookup_mntaxa}
\alias{lookup_mntaxa}
\title{Create MNTaxa look-up table}
\usage{
lookup_mntaxa(
  taxonomy_levels = FALSE,
  sources = FALSE,
  phys = FALSE,
  origin = FALSE,
  common = FALSE,
  cvals = FALSE,
  exclude = FALSE,
  replace_sub_var = FALSE,
  replace_family = FALSE,
  replace_genus = FALSE,
  drop_higher = FALSE,
  higher_include = c("Belonia", "Chara", "Lychnothamnus", "Nitella", "Nitellopsis",
    "Spirogyra", "Tolypella"),
  excluded_duplicates = FALSE,
  clean_duplicates = FALSE,
  group_accepted = FALSE
)
}
\arguments{
\item{taxonomy_levels}{Binary option (TRUE, FALSE) to include rank of taxa and taxonomic parents.}

\item{sources}{Binary option (TRUE, FALSE) to include authorities and publications.}

\item{phys}{Binary option (TRUE, FALSE) to include physiognomy.}

\item{origin}{Binary option (TRUE, FALSE) to include origin.}

\item{common}{Binary option (TRUE, FALSE) to include common names.}

\item{cvals}{Binary option (TRUE, FALSE) to include c-values.}

\item{exclude}{Binary option (TRUE, FALSE) to include information about excluding taxa from analyses.}

\item{replace_sub_var}{Binary option (TRUE, FALSE) to replace subspecies and varieties with their species when the species is present in MNTaxa.}

\item{replace_family}{Binary option (TRUE, FALSE) to replace families with species or genus when only one species or genus, respectively, is included in MNtaxa.}

\item{replace_genus}{Binary option (TRUE, FALSE) to replace genera with species when only one species is included in MNtaxa.}

\item{drop_higher}{Binary option (TRUE, FALSE) to drop taxonomic levels above species from the accepted taxa list. Implemented after \code{replace_family} and \code{replace_genus}. See \code{higher_include} for exceptions.}

\item{higher_include}{Vector of accepted genera or family names to be include in look-up table. The default are included because they don't have lower taxa in the accepted species list (Belonia, Chara, Lychnothamnus, Nitella, Nitellopsis, Spirogyra, Tolypella). Only implemented if \code{drop_higher} = TRUE..}

\item{excluded_duplicates}{Binary option (TRUE, FALSE) to reduce duplicate accepted name assignments by removing those with exclusions documented in MNTaxa.}

\item{clean_duplicates}{Binary option (TRUE, FALSE) to remove duplicate accepted name assignments by assigning a taxon to itself when available and assigning the most likely taxon for a handful of hand-curated duplicates. Recommended for one-to-one taxon-to-accepted look-up table.}

\item{group_accepted}{Binary option (TRUE, FALSE) to group all potential accepted names. Recommended for one-to-one taxon id-to-analysis group look-up table.}
}
\value{
Tibble of taxon names paired with accepted names and optional attributes. If both \code{clean_duplicates} and \code{group_accepted} are selected, \code{group_accepted} will be used.
}
\description{
Create MNTaxa look-up table
}
\examples{
lookup <- lookup_mntaxa()
}
