% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{origin_codes_raw}
\alias{origin_codes_raw}
\title{Native status codes table}
\format{
A data frame with the following columns:
\describe{
\item{id}{Integer. Primary key.}
\item{native_status}{Character. Description of the native status (e.g., native, introduced).}
}
}
\source{
MNTaxa database, plants.native_statuses table.
}
\usage{
origin_codes_raw
}
\description{
Native status codes table
}
\keyword{datasets}
