% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pars_raw}
\alias{pars_raw}
\title{Taxonomic parents table}
\format{
A data frame with the following columns:
\describe{
\item{taxon_id}{Integer. Foreign key linking to \code{taxa_raw}.}
\item{parent_id}{Integer. Foreign key linking to the parent taxon in \code{taxa_raw}.}
}
}
\source{
MNTaxa database, plants.parents_taxa table.
}
\usage{
pars_raw
}
\description{
Taxonomic parents table
}
\keyword{datasets}
