% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{phys_codes_raw}
\alias{phys_codes_raw}
\title{Physiognomy codes table}
\format{
A data frame with the following columns:
\describe{
\item{id}{Integer. Primary key.}
\item{physiognomy}{Character. Name of the physiognomy class (e.g., forb, shrub, tree).}
\item{physiognomy_code}{Character. Abbreviated code for the physiognomy class.}
}
}
\source{
MNTaxa database, plants.physiognomies table.
}
\usage{
phys_codes_raw
}
\description{
Physiognomy codes table
}
\keyword{datasets}
