% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pubs_raw}
\alias{pubs_raw}
\title{Publication codes table}
\format{
A data frame with the following columns:
\describe{
\item{id}{Integer. Primary key.}
\item{title}{Character. Title of the publication.}
}
}
\source{
MNTaxa database, plants.publications table.
}
\usage{
pubs_raw
}
\description{
Publication codes table
}
\keyword{datasets}
