% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{rank_raw}
\alias{rank_raw}
\title{Taxonomic rank codes table}
\format{
A data frame with the following columns:
\describe{
\item{id}{Integer. Primary key.}
\item{rank}{Character. Name of the taxonomic rank (e.g., species, genus, family).}
}
}
\source{
MNTaxa database, plants.ranks table.
}
\usage{
rank_raw
}
\description{
Taxonomic rank codes table
}
\keyword{datasets}
