% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releve_taxa_data.R
\docType{data}
\name{releve_taxa}
\alias{releve_taxa}
\title{Releve taxa that aren't in MNTaxa}
\format{
\subsection{\code{releve_taxa}}{

A data frame with 7 rows and 3 columns:
\describe{
\item{taxon_id}{ID that's based on releve data}
\item{taxon}{taxon name}
\item{syonymy_id}{ID created to include these taxa in accepted table}
}
}
}
\source{
MN DNR releves
}
\usage{
releve_taxa
}
\description{
Dataframe of taxon_id and taxon (name) to be used to include these taxa within MNTaxa output.
}
\keyword{datasets}
