% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{syn_comm_raw}
\alias{syn_comm_raw}
\title{Common names table}
\format{
A data frame with the following columns:
\describe{
\item{synonymy_id}{Integer. Foreign key linking to \code{syns_raw}.}
\item{common_name}{Character. Common name associated with the synonymy group.}
}
}
\source{
MNTaxa database, plants.common_names table.
}
\usage{
syn_comm_raw
}
\description{
Common names table
}
\keyword{datasets}
