% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{syn_cvals_raw}
\alias{syn_cvals_raw}
\title{Conservatism coefficients (C-values) table}
\format{
A data frame with the following columns:
\describe{
\item{synonymy_id}{Integer. Foreign key linking to \code{syns_raw}.}
\item{conservationism_coefficient}{Integer. C-value assigned to the taxon, ranging from 0 to 10.}
\item{end_date}{Date. Date the C-value assignment ended. \code{NA} if currently active.}
\item{publication_id}{Integer. Foreign key linking to \code{pubs_raw} indicating the source of the C-value assignment.}
}
}
\source{
MNTaxa database, plants.conservationism_coefficients table.
}
\usage{
syn_cvals_raw
}
\description{
Conservatism coefficients (C-values) table
}
\keyword{datasets}
