% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{syn_exclude_raw}
\alias{syn_exclude_raw}
\title{Synonymy-exclusion crosswalk table}
\format{
A data frame with the following columns:
\describe{
\item{synonymy_id}{Integer. Foreign key linking to \code{syns_raw}.}
\item{begin_date}{Date. Date the exclusion became active.}
\item{end_date}{Date. Date the exclusion ended. \code{NA} if currently active.}
\item{excluded_reason_id}{Integer. Foreign key linking to \code{exclude_codes_raw}.}
}
}
\source{
MNTaxa database, plants.excluded_reasons_synonymies table.
}
\usage{
syn_exclude_raw
}
\description{
Synonymy-exclusion crosswalk table
}
\keyword{datasets}
