% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{syn_or_raw}
\alias{syn_or_raw}
\title{Synonymy-origin crosswalk table}
\format{
A data frame with the following columns:
\describe{
\item{id}{Integer. Primary keyForeign key linking to synonymy_id in \code{syns_raw}.}
\item{native_status_id}{Integer. Foreign key linking to \code{origin_codes_raw}.}
}
}
\source{
MNTaxa database, plants.synonymies table.
}
\usage{
syn_or_raw
}
\description{
Synonymy-origin crosswalk table
}
\keyword{datasets}
