% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{syn_phys_raw}
\alias{syn_phys_raw}
\title{Synonymy-physiognomy crosswalk table}
\format{
A data frame with the following columns:
\describe{
\item{synonymy_id}{Integer. Foreign key linking to \code{syns_raw}.}
\item{physiognomy_id}{Integer. Foreign key linking to \code{phys_codes_raw}.}
\item{is_acceptable}{Integer. Binary indicator of whether this physiognomy assignment is acceptable (1) or not (0).}
}
}
\source{
MNTaxa database, plants.physiognomies_synonymies table.
}
\usage{
syn_phys_raw
}
\description{
Synonymy-physiognomy crosswalk table
}
\keyword{datasets}
