% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{syns_raw}
\alias{syns_raw}
\title{Synonymies and taxa table}
\format{
A data frame with the following columns:
\describe{
\item{taxon_id}{Integer. Foreign key linking to \code{taxa_raw}.}
\item{synonymy_id}{Integer. Identifier for the synonymy group.}
\item{d_list_beg_date}{Date. Date the taxon was added to the accepted list for this synonymy group.}
\item{d_list_end_date}{Date. Date the taxon was removed from the accepted list for this synonymy group. \code{NA} indicates currently accepted.}
}
}
\source{
MNTaxa database, plants.synonymies_taxa table.
}
\usage{
syns_raw
}
\description{
Synonymies and taxa table
}
\keyword{datasets}
