% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_mntaxa.R
\name{taxa_mntaxa}
\alias{taxa_mntaxa}
\title{Create table of all taxa}
\usage{
taxa_mntaxa(taxonomy_levels = FALSE, sources = FALSE)
}
\arguments{
\item{taxonomy_levels}{Binary option (TRUE, FALSE) to include rank of taxa and taxonomic parents.}

\item{sources}{Binary option (TRUE, FALSE) to include authorities and publications.}
}
\value{
Tibble of currently accepted taxa in MNTaxa and any additional information requested by the user.
}
\description{
Create table of all taxa
}
\examples{
taxa <- taxa_mntaxa()
}
