% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{taxa_raw}
\alias{taxa_raw}
\title{All taxa table}
\format{
A data frame with the following columns:
\describe{
\item{id}{Integer. Primary key.}
\item{taxon}{Character. Scientific name of the taxon.}
\item{ss_sl}{Character. Qualifier to scientific name.}
\item{is_hybrid}{Integer. Binary indicator of whether the taxon is a hybrid (1) or not (0).}
\item{hybrid_parents}{Character. Names of parent taxa for hybrids. \code{NA} for non-hybrids.}
\item{rank_id}{Integer. Foreign key linking to \code{rank_raw}.}
\item{author_id}{Integer. Foreign key linking to \code{auth_raw}.}
\item{publication_id}{Integer. Foreign key linking to \code{pubs_raw}.}
\item{begin_date}{Date. Date the taxon was added to MNTaxa.}
\item{end_date}{Date. Date the taxon was removed from MNTaxa. \code{NA} if current.}
\item{created_at}{POSIXct. Timestamp of record creation.}
\item{updated_at}{POSIXct. Timestamp of most recent record update.}
}
}
\source{
MNTaxa database, plants.taxa table.
}
\usage{
taxa_raw
}
\description{
All taxa table
}
\keyword{datasets}
