% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mntaxa.R
\name{load_mntaxa}
\alias{load_mntaxa}
\title{Load MNTaxa tables}
\usage{
load_mntaxa(
  synonymies = TRUE,
  all_taxa = TRUE,
  taxonomy_levels = FALSE,
  sources = FALSE,
  phys = FALSE,
  origin = FALSE,
  common = FALSE,
  cvals = FALSE,
  exclude = FALSE,
  envir = .GlobalEnv
)
}
\arguments{
\item{synonymies}{Binary option (TRUE, FALSE) to load synonymy table.}

\item{all_taxa}{Binary option (TRUE, FALSE) to load all taxa table.}

\item{taxonomy_levels}{Binary option (TRUE, FALSE) to load tables pertaining to taxonomy levels, including rank of taxa and taxonomic parents.}

\item{sources}{Binary option (TRUE, FALSE) to load tables for authorities and publications.}

\item{phys}{Binary option (TRUE, FALSE) to load physiognomy tables.}

\item{origin}{Binary option (TRUE, FALSE) to load origin tables.}

\item{common}{Binary option (TRUE, FALSE) to load common names table.}

\item{cvals}{Binary option (TRUE, FALSE) to load c-value table.}

\item{exclude}{Binary option (TRUE, FALSE) to load tables containing information about excluding taxa from analyses.}

\item{envir}{The environment where tables will be loaded (default is user's global environment).}
}
\value{
Requested tables
}
\description{
\code{load_mntaxa()}loads a selection of MNTaxa tables directly from the database into the specified environment.
}
\examples{
load_mntaxa()
}
