% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_mntaxa.R
\name{lookup_mntaxa}
\alias{lookup_mntaxa}
\title{Title}
\usage{
lookup_mntaxa(
  taxonomy_levels = FALSE,
  sources = FALSE,
  phys = FALSE,
  origin = FALSE,
  common = FALSE,
  cvals = FALSE,
  exclude = FALSE,
  replace_sub_var = FALSE,
  replace_family = FALSE,
  replace_genus = FALSE,
  excluded_duplicates = FALSE,
  clean_duplicates = FALSE,
  group_accepted = FALSE
)
}
\arguments{
\item{taxonomy_levels}{Binary option (TRUE, FALSE) to include rank of taxa and taxonomic parents.}

\item{sources}{Binary option (TRUE, FALSE) to include authorities and publications.}

\item{phys}{Binary option (TRUE, FALSE) to include physiognomy.}

\item{origin}{Binary option (TRUE, FALSE) to include origin.}

\item{common}{Binary option (TRUE, FALSE) to include common names.}

\item{cvals}{Binary option (TRUE, FALSE) to include c-values.}

\item{exclude}{Binary option (TRUE, FALSE) to include information about excluding taxa from analyses.}

\item{replace_sub_var}{Binary option (TRUE, FALSE) to replace subspecies and varieties with their species when the species is present in MNTaxa.}

\item{replace_family}{Binary option (TRUE, FALSE) to replace families with species or genus when only one species or genus, respectively, is included in MNtaxa.}

\item{replace_genus}{Binary option (TRUE, FALSE) to replace genera with species when only one species is included in MNtaxa.}

\item{excluded_duplicates}{Binary option (TRUE, FALSE) to reduce duplicate accepted name assignments by removing those with exclusions documented in MNTaxa.}

\item{clean_duplicates}{Binary option (TRUE, FALSE) to remove duplicate accepted name assignments by assigning a taxon to itself when available, assigning a subspecies/variety to the species when available, and assigning the most likely taxon for a handful of hand-curated duplicates.}

\item{group_accepted}{Binary option (TRUE, FALSE) to group all potential accepted names.}
}
\value{
Tibble of taxon names paired with accepted names and optional attributes.
}
\description{
Title
}
\examples{
lookup <- lookup_mntaxa()
}
