% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subvar_mntaxa.R
\name{subvar_mntaxa}
\alias{subvar_mntaxa}
\title{Replace accepted subspecies and varieties with species}
\usage{
subvar_mntaxa()
}
\value{
Tibble with acc_taxon_id corresponding to the subspecies/variety and columns for accepted taxon attributes associated with the corresponding species and ending with "_rep" so that they can be used to replace subspecies/variety attributes.
}
\description{
Replace accepted subspecies and varieties with species
}
\examples{
acc_sub_var <- subvar_mntaxa()
}
